refactor(logging): enforce single responsibility file structure for torrent payloads
