+++ qBitControl/Logging/String+Logging.swift
@@ -1,0 +1,23 @@
+//
+//  String+Logging.swift
+//  qBitControl
+//
+
+import Foundation
+import CryptoKit
+
+extension String {
+    func sha256Hash() -> String {
+        let inputData = Data(self.utf8)
+        let hashed = SHA256.hash(data: inputData)
+        return hashed.compactMap { String(format: "%02x", $0) }.joined()
+    }
+    
+    func replacingHomeDirectoryWithTilde() -> String {
+        let home = NSHomeDirectory()
+        if self.hasPrefix(home) {
+            return "~" + self.dropFirst(home.count)
+        }
+        return self
+    }
+}
